Separate webserver.py into module and script files (quick fix to #10).
